Finalize typings
