+++ packages/api/.gitignore
@@ -1,3 +1,2 @@
 /lib
 /node_modules
-/types
