+++ packages/bundle/.gitignore
@@ -2,4 +2,3 @@
 /lib
 /lib.instrumented
 /node_modules
-/types
