+++ packages/bundle/src/adaptiveCards/AdaptiveCardsStyleOptions.ts
@@ -1,0 +1,23 @@
+type AdaptiveCardStyleOptions = {
+  /**
+   * Cards styling
+   */
+
+  // TODO: Put JSDoc on every options.
+  cardEmphasisBackgroundColor?: string;
+
+  /**
+   * Cards: Adaptive Card push button
+   */
+
+  cardPushButtonBackgroundColor?: string;
+  cardPushButtonTextColor?: string;
+
+  /**
+   * Cards: Rich Cards
+   * Enable title (and subtitle) wrapping
+   */
+  richCardWrapTitle?: boolean;
+};
+
+export default AdaptiveCardStyleOptions;
