+++ packages/bundle/src/index.ts
@@ -1,3 +1,5 @@
+// DOC: To export anything from this file, add it to index.tsx, which is the .d.ts for this file.
+
 /* eslint dot-notation: ["error", { "allowPattern": "^WebChat$" }] */
 // window['WebChat'] is required for TypeScript
