+++ packages/bundle/src/tsconfig.json
@@ -2,15 +2,14 @@
   "compilerOptions": {
     "allowSyntheticDefaultImports": true,
     "declaration": true,
-    "declarationDir": "../types",
+    "declarationDir": "../lib",
     "declarationMap": true,
     "emitDeclarationOnly": true,
-    // "isolatedModules": true,
     "jsx": "react",
     "preserveWatchOutput": true,
     "pretty": true,
     "skipLibCheck": true,
     "sourceMap": true
   },
-  "files": ["FullBundleStyleOptions.ts", "types.ts"]
+  "files": ["index.tsx", "index-es5.tsx", "index-minimal.tsx"]
 }
