--- packages/bundle/src/types.ts
@@ -1,7 +1,0 @@
-// Any changes to this file, please include it in tsconfig.json.
-
-import { StyleOptions as MinimalStyleOptions } from 'botframework-webchat-api';
-
-import FullBundleStyleOptions from './FullBundleStyleOptions';
-
-export type { FullBundleStyleOptions, MinimalStyleOptions };
