+++ packages/isomorphic-react-dom/package.json
@@ -8,7 +8,6 @@
     "build": "npm run build:babel && npm run build:webpack",
     "build:babel": "babel src --out-dir lib --verbose",
     "build:webpack": "webpack-cli",
-    "eslint": "eslint src/**/*.js --fix",
     "start": "npm run start:note && npm run start:babel && npm run start:webpack",
     "start:babel": "npm run build:babel",
     "start:note": "echo \"isomorphic-react-dom is not under active development; running in non-watch mode.\"",
